+++ package/Editor/ContentPipeline.cs
@@ -74,7 +74,7 @@ namespace UnityEditor.Build.Pipeline
             ReturnCode exitCode;
             result = new BundleBuildResults();
 
-            using (new SceneStateCleanup())
+        //    using (new SceneStateCleanup())
             using (new BuildInterfacesWrapper())
             using (var progressTracker = new ProgressTracker())
             using (var buildCache = new BuildCache(parameters.CacheServerHost, parameters.CacheServerPort))