+++ package/Editor/Shared/BuildSpriteData.cs
@@ -1,11 +1,13 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using UnityEditor.Build.Pipeline.Interfaces;
 
 namespace UnityEditor.Build.Pipeline
 {
+    [Serializable]
     public class BuildSpriteData : IBuildSpriteData
     {
-        public Dictionary<GUID, SpriteImporterData> ImporterData { get; private set; }
+        public Dictionary<GUID, SpriteImporterData> ImporterData { get; set; }
 
         public BuildSpriteData()
         {