+++ package/Editor/Shared/BundleExplictObjectLayout.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using UnityEditor.Build.Content;
 using UnityEditor.Build.Pipeline.Interfaces;
 
@@ -7,10 +8,11 @@ namespace UnityEditor.Build.Pipeline
     /// <summary>
     /// Optional context object used for overriding the location where specific objects will be serialized
     /// </summary>
+    [Serializable]
     public class BundleExplictObjectLayout : IBundleExplictObjectLayout
     {
         /// <inheritdoc />
-        public Dictionary<ObjectIdentifier, string> ExplicitObjectLocation { get; private set; }
+        public Dictionary<ObjectIdentifier, string> ExplicitObjectLocation { get; set; }
 
         /// <summary>
         /// Default constructor, initializes properties to defaults