+++ package/Editor/Utilities/BuildCache.cs
@@ -18,8 +18,13 @@ namespace UnityEditor.Build.Pipeline.Utilities
     {
         const string k_CachePath = "Library/BuildCache";
 
-        Dictionary<GUID, CacheEntry> m_GuidToHash = new Dictionary<GUID, CacheEntry>();
-        Dictionary<string, CacheEntry> m_PathToHash = new Dictionary<string, CacheEntry>();
+        Dictionary<KeyValuePair<GUID, int>, CacheEntry> m_GuidToHash = new Dictionary<KeyValuePair<GUID, int>, CacheEntry>();
+        Dictionary<KeyValuePair<string, int>, CacheEntry> m_PathToHash = new Dictionary<KeyValuePair<string, int>, CacheEntry>();
+
+        Thread m_ActiveWriteThread;
+
+        [NonSerialized]
+        Hash128 m_GlobalHash;
 
         [NonSerialized]
         CacheServerUploader m_Uploader;
@@ -27,10 +32,15 @@ namespace UnityEditor.Build.Pipeline.Utilities
         [NonSerialized]
         CacheServerDownloader m_Downloader;
 
-        public BuildCache() { }
+        public BuildCache()
+        {
+            m_GlobalHash = CalculateGlobalArtifactVersionHash();
+        }
 
         public BuildCache(string host, int port = 8126)
         {
+            m_GlobalHash = CalculateGlobalArtifactVersionHash();
+
             if (string.IsNullOrEmpty(host))
                 return;
 
@@ -38,8 +48,30 @@ namespace UnityEditor.Build.Pipeline.Utilities
             m_Downloader = new CacheServerDownloader(this, host, port);
         }
 
+        // internal for testing purposes only
+        internal void OverrideGlobalHash(Hash128 hash)
+        {
+            m_GlobalHash = hash;
+            if (m_Uploader != null)
+                m_Uploader.SetGlobalHash(m_GlobalHash);
+            if (m_Downloader != null)
+                m_Downloader.SetGlobalHash(m_GlobalHash);
+        }
+
+        static Hash128 CalculateGlobalArtifactVersionHash()
+        {
+            return HashingMethods.Calculate(PlayerSettings.scriptingRuntimeVersion, Application.unityVersion).ToHash128();
+        }
+
+        internal void ClearCacheEntryMaps()
+        {
+            m_GuidToHash.Clear();
+            m_PathToHash.Clear();
+        }
+
         public void Dispose()
         {
+            SyncPendingSaves();
             if (m_Uploader != null)
                 m_Uploader.Dispose();
             if (m_Downloader != null)
@@ -49,13 +81,14 @@ namespace UnityEditor.Build.Pipeline.Utilities
         }
 
         /// <inheritdoc />
-        public CacheEntry GetCacheEntry(GUID asset)
+        public CacheEntry GetCacheEntry(GUID asset, int version = 1)
         {
             CacheEntry entry;
-            if (m_GuidToHash.TryGetValue(asset, out entry))
+            KeyValuePair<GUID, int> key = new KeyValuePair<GUID, int>(asset, version);
+            if (m_GuidToHash.TryGetValue(key, out entry))
                 return entry;
 
-            entry = new CacheEntry { Guid = asset };
+            entry = new CacheEntry { Guid = asset, Version = version };
             string path = AssetDatabase.GUIDToAssetPath(asset.ToString());
             entry.Type = CacheEntry.EntryType.Asset;
 
@@ -68,45 +101,48 @@ namespace UnityEditor.Build.Pipeline.Utilities
                     entry.Hash = HashingMethods.CalculateFile(path).ToHash128();
             }
 
-            m_GuidToHash[entry.Guid] = entry;
+            entry.Hash = HashingMethods.Calculate(entry.Hash, entry.Version).ToHash128();
+
+            m_GuidToHash[key] = entry;
             return entry;
         }
 
         /// <inheritdoc />
-        public CacheEntry GetCacheEntry(string path)
+        public CacheEntry GetCacheEntry(string path, int version = 1)
         {
             CacheEntry entry;
-            if (m_PathToHash.TryGetValue(path, out entry))
+            KeyValuePair<string, int> key = new KeyValuePair<string, int>(path, version);
+            if (m_PathToHash.TryGetValue(key, out entry))
                 return entry;
 
-            entry = new CacheEntry { File = path };
+            entry = new CacheEntry { File = path, Version = version };
             entry.Guid = HashingMethods.Calculate("FileHash", entry.File).ToGUID();
-            entry.Hash = HashingMethods.CalculateFile(entry.File).ToHash128();
+            entry.Hash = HashingMethods.Calculate(HashingMethods.CalculateFile(entry.File), entry.Version).ToHash128();
             entry.Type = CacheEntry.EntryType.File;
 
-            m_PathToHash[entry.File] = entry;
+            m_PathToHash[key] = entry;
             return entry;
         }
 
         /// <inheritdoc />
-        public CacheEntry GetCacheEntry(ObjectIdentifier objectID)
+        public CacheEntry GetCacheEntry(ObjectIdentifier objectID, int version = 1)
         {
             if (objectID.guid.Empty())
-                return GetCacheEntry(objectID.filePath);
-            return GetCacheEntry(objectID.guid);
+                return GetCacheEntry(objectID.filePath, version);
+            return GetCacheEntry(objectID.guid, version);
         }
 
-        CacheEntry GetUpdatedCacheEntry(CacheEntry entry)
+        internal CacheEntry GetUpdatedCacheEntry(CacheEntry entry)
         {
             if (entry.Type == CacheEntry.EntryType.File)
-                return GetCacheEntry(entry.File);
+                return GetCacheEntry(entry.File, entry.Version);
             if (entry.Type == CacheEntry.EntryType.Asset)
-                return GetCacheEntry(entry.Guid);
+                return GetCacheEntry(entry.Guid, entry.Version);
             return entry;
         }
 
         /// <inheritdoc />
-        public bool NeedsRebuild(CachedInfo info)
+        public bool HasAssetOrDependencyChanged(CachedInfo info)
         {
             if (info == null || !info.Asset.IsValid() || info.Asset != GetUpdatedCacheEntry(info.Asset))
                 return true;
@@ -124,14 +160,16 @@ namespace UnityEditor.Build.Pipeline.Utilities
         public string GetCachedInfoFile(CacheEntry entry)
         {
             var guid = entry.Guid.ToString();
-            return string.Format("{0}/{1}/{2}/{2}_{3}.info", k_CachePath, guid.Substring(0, 2), guid, entry.Hash.ToString());
+            string finalHash = HashingMethods.Calculate(m_GlobalHash, entry.Hash).ToString();
+            return string.Format("{0}/{1}/{2}/{2}_{3}.info", k_CachePath, guid.Substring(0, 2), guid, finalHash);
         }
 
         /// <inheritdoc />
         public string GetCachedArtifactsDirectory(CacheEntry entry)
         {
             var guid = entry.Guid.ToString();
-            return string.Format("{0}/{1}/{2}/{3}", k_CachePath, guid.Substring(0, 2), guid, entry.Hash.ToString());
+            string finalHash = HashingMethods.Calculate(m_GlobalHash, entry.Hash).ToString();
+            return string.Format("{0}/{1}/{2}", k_CachePath, guid.Substring(0, 2), finalHash);
         }
 
         class FileOperations
@@ -250,12 +288,13 @@ namespace UnityEditor.Build.Pipeline.Utilities
                 }
                 cachedInfos.Add(info);
             }
+            thread.Join();
             ((IDisposable)ops.waitLock).Dispose();
 
             // Validate cached data is reusable
             for (int i = 0; i < cachedInfos.Count; i++)
             {
-                if (NeedsRebuild(cachedInfos[i]))
+                if (HasAssetOrDependencyChanged(cachedInfos[i]))
                     cachedInfos[i] = null;
             }
 
@@ -282,8 +321,9 @@ namespace UnityEditor.Build.Pipeline.Utilities
             }
 
             // Start writing thread
-            Thread thread = new Thread(Write);
-            thread.Start(ops);
+            SyncPendingSaves();
+            m_ActiveWriteThread = new Thread(Write);
+            m_ActiveWriteThread.Start(ops);
 
             // Serialize data as previous data is being written out
             var formatter = new BinaryFormatter();
@@ -311,10 +351,24 @@ namespace UnityEditor.Build.Pipeline.Utilities
             }
         }
 
+        internal void SyncPendingSaves()
+        {
+            if(m_ActiveWriteThread != null)
+            {
+                m_ActiveWriteThread.Join();
+                m_ActiveWriteThread = null;
+            }
+        }
+
         [MenuItem("Window/Asset Management/Purge Build Cache", priority = 10)]
         public static void PurgeCache()
         {
-            if (!EditorUtility.DisplayDialog("Purge Build Cache", "Do you really want to purge your entire build cache?", "Yes", "No"))
+            PurgeCache(false);
+        }
+
+        public static void PurgeCache(bool prompt)
+        {
+            if (prompt && !EditorUtility.DisplayDialog("Purge Build Cache", "Do you really want to purge your entire build cache?", "Yes", "No"))
                 return;
 
             if (Directory.Exists(k_CachePath))